+++ main.go
@@ -60,7 +60,7 @@ func main() {
 	// Get Hypercloud Operating Mode!!!
 	hcMode := os.Getenv("HC_MODE")
 	util.TokenExpiredDate = os.Getenv("INVITATION_TOKEN_EXPIRED_DATE")
-
+	kafkaConsumer.KafkaGroupId = os.Getenv("KAFKA_GROUP_ID")
 	util.ReadFile()
 	caller.UpdateAuditResourceList()
 
@@ -107,7 +107,7 @@ func main() {
 
 	// Metering Cron Job
 	cronJob.AddFunc("0 */1 * ? * *", metering.MeteringJob)
-	cronJob.AddFunc("@hourly", audit.UpdateAuditResource)
+	// cronJob.AddFunc("@hourly", audit.UpdateAuditResource)
 	cronJob.Start()
 
 	// Hyperauth Event Consumer
@@ -154,6 +154,7 @@ func main() {
 	mux.HandleFunc("/audit", serveAudit)
 	mux.HandleFunc("/audit/batch", serveAuditBatch)
 	mux.HandleFunc("/audit/resources", serveAuditResources)
+	mux.HandleFunc("/audit/verb", serveAuditVerb)
 	mux.HandleFunc("/audit/websocket", serveAuditWss)
 	mux.HandleFunc("/inject/pod", serveSidecarInjectionForPod)
 	mux.HandleFunc("/inject/deployment", serveSidecarInjectionForDeploy)
@@ -510,6 +511,16 @@ func serveAudit(w http.ResponseWriter, r *http.Request) {
 		audit.AddAudit(w, r)
 	case http.MethodPut:
 	case http.MethodDelete:
+	default:
+		//error
+	}
+}
+
+func serveAuditVerb(w http.ResponseWriter, r *http.Request) {
+	klog.Infof("Http request: method=%s, uri=%s", r.Method, r.URL.Path)
+	switch r.Method {
+	case http.MethodGet:
+		audit.ListAuditVerb(w, r)
 	default:
 		//error
 	}