+++ CHANGELOG.md
@@ -1,6 +1,49 @@
 # Hypercloud-api-server changelog!!
 All notable changes to this project will be documented in this file.
 
+<!-------------------- v5.0.18.0 start -------------------->
+
+## Hypercloud-api-server 5.0.18.0 (2021. 07. 09. (금) 09:36:55 KST)
+
+### Added
+  - [feat] audit verb list api 추가 by chosangwon93
+
+### Changed
+
+### Fixed
+
+### CRD yaml
+
+### Etc
+
+<!-------------------- v5.0.17.5 start -------------------->
+
+## Hypercloud-api-server 5.0.17.5 (2021. 07. 08. (목) 13:29:58 KST)
+
+### Added
+
+### Changed
+
+### Fixed
+
+### CRD yaml
+
+### Etc
+
+<!-------------------- v5.0.17.3 start -------------------->
+
+## Hypercloud-api-server 5.0.17.3 (2021. 07. 07. (수) 18:06:51 KST)
+
+### Added
+
+### Changed
+
+### Fixed
+
+### CRD yaml
+
+### Etc
+
 <!-------------------- v5.0.17.2 start -------------------->
 
 ## Hypercloud-api-server 5.0.17.2 (2021. 07. 07. (수) 17:31:17 KST)